Refactor post app
